requirements.txt
Flask==2.2.5
Pillow==9.4.0
torch==1.11.0+cpu
torchvision==0.12.0+cpu
numpy==1.21.5